Pin refactor to move Input and Output pin's to their own source files
