+++ TACT/src/InputPin.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "Pin.h"
+
+class InputPin : public Pin {
+public:
+	InputPin(int id);
+	InputPin(int id, PinType type);
+private:
+	void StartAttribute() override;
+	void EndAttribute() override;
+};
